* Move configs into constants. * Use filename as subdir (group frames together)

+++ video_upscale_cpu.py
@@ -1,5 +1,6 @@
 from basicsr.archs.rrdbnet_arch import RRDBNet
 from dotenv import load_dotenv
+from pathlib import Path
 from realesrgan import RealESRGANer
 from tqdm import tqdm
 import cv2
@@ -15,23 +16,24 @@ VIDEO_OUTPUT = os.getenv('VIDEO_OUTPUT')
 TEMP_DIR = os.getenv('TEMP_DIR')
 ESRGAN_MODEL_PATH = os.getenv('ESRGAN_MODEL_PATH')
 
-DIR_FRAMES_SOURCE = "_frames_extracted"
-DIR_FRAMES_UPSCALE = "_frames_upscale"
-SCALE_FACTOR = 2 # "4" also available
+DIR_FRAMES_SOURCE = "source"
+DIR_FRAMES_UPSCALE = "upscaled"
+SCALE_FACTOR = 4 # "4" also available
 VIDEO_FPS = 30 # TODO: we should use the video's FPS and not a constant
+DONT_OVERFLOW_RAM = False
+IMAGES_OUTPUT_FORMAT = "png"
 
 
 
 def extract_frames(video_path: str, dir_output: str) -> None:
     os.makedirs(dir_output, exist_ok = True)
     subprocess.run([
-        "ffmpeg", "-i", video_path, f"{dir_output}/frame_%05d.png"
+        "ffmpeg", "-i", video_path, f"{dir_output}/frame_%05d.{IMAGES_OUTPUT_FORMAT}"
     ], check = True)
 
 
 
 def _get_upscaler_model(model_path: str, scale_factor: int) -> RealESRGANer:
-    DONT_OVERFLOW_RAM = True
     upscaler_model = RRDBNet(
         num_in_ch = 3,
         num_out_ch = 3,
@@ -54,7 +56,10 @@ def _get_upscaler_model(model_path: str, scale_factor: int) -> RealESRGANer:
 
 
 
-def _upscale_img(model: RealESRGANer, path_input: str, path_output: str, scale_factor: int):
+def _upscale_img(model: RealESRGANer, path_input: str, path_output: str, scale_factor: int) -> None:
+    if os.path.exists(path_output):
+        print(f"  upscale: skipping {path_output} (file already it exists)")
+        return
     img_source = cv2.imread(path_input)
     img_upscaled, _ = model.enhance(img_source, outscale = scale_factor)
     cv2.imwrite(path_output, img_upscaled)
@@ -64,16 +69,15 @@ def _upscale_img(model: RealESRGANer, path_input: str, path_output: str, scale_f
 def upscale_frames(dir_input: str, dir_output: str, model_path: str, scale_factor: int) -> None:
     os.makedirs(dir_output, exist_ok = True)
     model = _get_upscaler_model(model_path, scale_factor)
-    FORMAT = ".png" # TODO: this should be ".png", currently a hack to reduce sample for testing
+    # TODO: this should be ".png".
+    # Using "00.png" as a hack to reduce sample for testing purposes
+    end_of_path_format = f"00.{IMAGES_OUTPUT_FORMAT}"
     images = sorted(
-        f for f in os.listdir(dir_input) if f.endswith(FORMAT)
+        f for f in os.listdir(dir_input) if f.endswith(end_of_path_format)
     ) 
     for img_name in tqdm(images, desc = "Upscaling frames"):
         path_input = os.path.join(dir_input, img_name)
         path_output = os.path.join(dir_output, img_name)
-        if os.path.exists(path_output):
-            print(f"  skipping {path_output}, it exists")
-            continue
         _upscale_img(model, path_input, path_output, scale_factor)
 
 
@@ -88,18 +92,20 @@ def frames_to_video(frame_dir: str, output_path: str, fps: int) -> None:
 
 
 if __name__ == "__main__":
-    #extract_frames(
-    #    VIDEO_INPUT,
-    #    f"{TEMP_DIR}/{DIR_FRAMES_SOURCE}"
-    #)
+    file_name = Path(VIDEO_INPUT).stem
+    formatted_temp_dir = f"{TEMP_DIR}/{file_name} x{SCALE_FACTOR}"
+    extract_frames(
+        VIDEO_INPUT,
+        f"{formatted_temp_dir}/{DIR_FRAMES_SOURCE}"
+    )
     upscale_frames(
-        f"{TEMP_DIR}/{DIR_FRAMES_SOURCE}",
-        f"{TEMP_DIR}/{DIR_FRAMES_UPSCALE}",
+        f"{formatted_temp_dir}/{DIR_FRAMES_SOURCE}",
+        f"{formatted_temp_dir}/{DIR_FRAMES_UPSCALE}",
         ESRGAN_MODEL_PATH,
         SCALE_FACTOR
     )
-    #frames_to_video(
-    #    f"{TEMP_DIR}/{DIR_FRAMES_UPSCALE}",
-    #    VIDEO_OUTPUT,
-    #    1 #VIDEO_FPS
-    #)
+    frames_to_video(
+        f"{formatted_temp_dir}/{DIR_FRAMES_UPSCALE}",
+        VIDEO_OUTPUT,
+        VIDEO_FPS
+    )
